+++ defaults.py
@@ -348,6 +348,7 @@ class FlatCAMDefaults:
         "tools_mill_polish_margin": 0.0,
         "tools_mill_polish_overlap": 5,
         "tools_mill_polish_method": 0,
+
         "tools_mill_milling_type": 'both',
         "tools_mill_optimization_type": 'R',
         "tools_mill_search_time": 3,